add feature of  online ddl and supplement test case

+++ jimsql/jim-sdk/jim-core/src/main/java/io/jimdb/core/expression/ConditionChecker.java
@@ -20,9 +20,9 @@ package io.jimdb.core.expression;
  */
 public class ConditionChecker {
 
-  private long columnId;
+  private int columnId;
 
-  public ConditionChecker(long columnId) {
+  public ConditionChecker(int columnId) {
     this.columnId = columnId;
   }
